Make certain sections dependent on config
I've made the speakers/schedule section dependent on a config value per conference. That way the conference can decide when the speakers/schedule section are visible. Given that CFS and the speakers/schedule are fairly mutually exclusive it makes little sense to show them all at once. Also this allows for CFS to move higher up the page for people to notice.